# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hack-Flow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /hackathons:
    get:
      operationId: listHackathons
      tags: [hackathons]
      summary: List hackathons
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Hackathon list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HackathonListResponse"
    post:
      operationId: createHackathon
      tags: [hackathons]
      summary: Create hackathon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HackathonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
  /hackathons/{id}:
    get:
      operationId: getHackathon
      tags: [hackathons]
      summary: Get hackathon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hackathon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateHackathon
      tags: [hackathons]
      summary: Update hackathon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HackathonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hackathon"
    delete:
      operationId: deleteHackathon
      tags: [hackathons]
      summary: Delete hackathon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /hackathons/{id}/register:
    post:
      operationId: registerForHackathon
      tags: [hackathons]
      summary: Register current user for hackathon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /hackathons/{id}/assign-role:
    post:
      operationId: assignHackathonRole
      tags: [hackathons]
      summary: Assign mentor or jury to a hackathon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRoleInput"
      responses:
        "200":
          description: Role assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /hackathons/{id}/milestones:
    get:
      operationId: getHackathonMilestones
      tags: [hackathons]
      summary: Get hackathon milestones/timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Milestones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [hackathons]
      summary: Create milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /hackathons/{id}/criteria:
    get:
      operationId: getEvaluationCriteria
      tags: [evaluation]
      summary: Get evaluation criteria for a hackathon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Criteria list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvaluationCriterion"
    post:
      operationId: createEvaluationCriterion
      tags: [evaluation]
      summary: Create evaluation criterion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationCriterionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationCriterion"

  # TEAMS
...
  /chat/rooms/{roomId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: Get messages in a chat room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message in a chat room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # ANALYTICS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # AUTH
    RegisterInput:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        role:
          type: string
          enum: [participant, organizer, mentor, jury, admin]
        skills:
          type: array
          items:
            type: string
        bio:
          type: string
        timezone:
          type: string
        githubUrl:
          type: string
        linkedinUrl:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, accessToken]
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
    User:
      type: object
      required: [id, email, name, role, isActive, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [participant, organizer, mentor, jury, admin]
        isActive:
          type: boolean
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        timezone:
          type: ["string", "null"]
        githubUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Hackathon:
      type: object
      required: [id, title, description, status, startDate, endDate, maxTeamSize, minTeamSize, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, open, active, judging, completed, cancelled]
        bannerUrl:
          type: ["string", "null"]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        registrationDeadline:
          type: ["string", "null"]
          format: date-time
        maxTeamSize:
          type: integer
        minTeamSize:
          type: integer
        maxParticipants:
          type: ["integer", "null"]
        prizePool:
          type: ["string", "null"]
        techStack:
          type: array
          items:
            type: string
        participantCount:
          type: integer
        teamCount:
          type: integer
        projectCount:
          type: integer
        organizerId:
          type: integer
        createdAt:
          type: string
          format: date-time
    HackathonInput:
      type: object
      required: [title, description, startDate, endDate, maxTeamSize, minTeamSize]
      properties:
        title:
          type: string
        description:
          type: string
        bannerUrl:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        maxTeamSize:
          type: integer
        minTeamSize:
          type: integer
        maxParticipants:
          type: integer
        prizePool:
          type: string
        techStack:
          type: array
          items:
            type: string
    HackathonUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, open, active, judging, completed, cancelled]
        bannerUrl:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        registrationDeadline:
          type: string
          format: date-time
        maxTeamSize:
          type: integer
        minTeamSize:
          type: integer
        maxParticipants:
          type: integer
        prizePool:
          type: string
        techStack:
          type: array
          items:
            type: string
    HackathonListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Hackathon"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AssignRoleInput:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: integer
        role:
          type: string
          enum: [mentor, jury]

    # MILESTONES
    Milestone:
      type: object
      required: [id, hackathonId, title, dueDate, createdAt]
      properties:
        id:
          type: integer
        hackathonId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date-time
        isCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MilestoneInput:
      type: object
      required: [title, dueDate]
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time

    # TEAMS
...
    EvaluationCriterion:
      type: object
      required: [id, hackathonId, name, weight, maxScore]
      properties:
        id:
          type: integer
        hackathonId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        weight:
          type: number
          format: float
        maxScore:
          type: integer
    EvaluationCriterionInput:
      type: object
      required: [name, weight, maxScore]
      properties:
        name:
          type: string
        description:
          type: string
        weight:
          type: number
          format: float
        maxScore:
          type: integer
...
    ChatMessage:
      type: object
      required: [id, roomId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        senderId:
          type: integer
        senderName:
          type: string
        senderAvatar:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ANALYTICS
...