# lib/api-zod/src/generated/types/hackathonInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hack-Flow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HackathonInput {
  title: string;
  description: string;
  bannerUrl?: string;
  startDate: Date;
  endDate: Date;
  registrationDeadline?: Date;
  maxTeamSize: number;
  minTeamSize: number;
  maxParticipants?: number;
  prizePool?: string;
  techStack?: string[];
}