# lib/api-zod/src/generated/types/hackathonListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hack-Flow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Hackathon } from './hackathon';

export interface HackathonListResponse {
  data: Hackathon[];
  total: number;
  page: number;
  limit: number;
}