# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Hack-Flow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  isActive: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  skills?: string[];
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  createdAt: Date;
}